pyproject.toml
[project]
name = "test-fast-api"
version = "0.1.0"
description = ""
authors = [
    {name = "fedor",email = "fedorlastv@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[all] (>=0.115.11,<0.116.0)",
    "uvicorn[all] (>=0.34.0,<0.35.0)",
    "black (>=25.1.0,<26.0.0)",
    "dishka (>=1.5.0,<2.0.0)",
    "sqlalchemy[all] (>=2.0.39,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "alembic[async] (>=1.15.1,<2.0.0)",
    "pyjwt[crypto] (>=2.10.1,<3.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
